Refactored example to remove duplicated code between REST and GRPC
