--- sdks/go/examples/oneOfEverything/grpc/static/catena.txt
@@ -1,1 +1,0 @@
-This is a txt file used to demonstrate the functionality of the getAsset endpoint.
