+++ sdks/go/examples/oneOfEverything/static/catena.txt
